+++ service/src/main/java/uk/nhs/adaptors/gp2gp/ehr/utils/SupportingInfoResourceExtractor.java
@@ -4,6 +4,7 @@ import java.util.Optional;
 
 import ca.uhn.fhir.model.api.TemporalPrecisionEnum;
 
+import org.apache.commons.lang3.StringUtils;
 import org.hl7.fhir.dstu3.model.BaseDateTimeType;
 import org.hl7.fhir.dstu3.model.DocumentReference;
 import org.hl7.fhir.dstu3.model.DiagnosticReport;
@@ -17,154 +18,196 @@ import org.hl7.fhir.dstu3.model.ReferralRequest;
 import uk.nhs.adaptors.gp2gp.ehr.mapper.MessageContext;
 
 public class SupportingInfoResourceExtractor {
-    private static final String CODE_SYSTEM = "2.16.840.1.113883.2.1.4.5.5";
+    private static final String PMIP_CODE_SYSTEM = "2.16.840.1.113883.2.1.4.5.5";
+    public static final String URN_OID_PREFIX = "urn:oid:";
 
     public static String extractDocumentReference(MessageContext messageContext, Reference reference) {
-        DocumentReference documentReference = messageContext
+        var documentReferenceOptional = messageContext
                 .getInputBundleHolder()
                 .getResource(reference.getReferenceElement())
-                .map(DocumentReference.class::cast)
-                .get();
+                .map(DocumentReference.class::cast);
 
-        StringBuilder stringBuilder = new StringBuilder();
+        if (documentReferenceOptional.isEmpty()) {
+            return "";
+        }
+
+        var stringBuilder = new StringBuilder();
         stringBuilder.append("{ Document:");
 
+        var documentReference = documentReferenceOptional.get();
+
         if (documentReference.hasCreated()) {
-            stringBuilder.append(" " + formatDateUsingDayPrecision(documentReference.getCreatedElement()));
+            stringBuilder
+                .append(" ")
+                .append(formatDateUsingDayPrecision(documentReference.getCreatedElement()));
         } else {
-            stringBuilder.append(" " + formatDateUsingDayPrecision(documentReference.getIndexedElement()));
+            stringBuilder
+                .append(" ")
+                .append(formatDateUsingDayPrecision(documentReference.getIndexedElement()));
         }
 
-        if (documentReference.hasType()) {
-            if (documentReference.getType().hasText()) {
-                stringBuilder.append(" " + documentReference.getType().getText());
-            } else if (documentReference.getType().hasCoding()) {
-                CodeableConceptMappingUtils.extractTextOrCoding(
-                    documentReference.getType()).ifPresent(docType -> stringBuilder.append(" " + docType)
-                );
-            }
-        }
+        CodeableConceptMappingUtils.extractTextOrCoding(documentReference.getType()).ifPresent(
+            code -> stringBuilder
+                .append(" ")
+                .append(code)
+        );
 
         if (documentReference.hasDescription()) {
-            stringBuilder.append(" " + documentReference.getDescription());
+            stringBuilder
+                .append(" ")
+                .append(documentReference.getDescription());
         }
-
         stringBuilder.append(" }");
 
         return stringBuilder.toString();
     }
 
     public static String extractObservation(MessageContext messageContext, Reference reference) {
-        Observation observation = messageContext
+        var observationOptional = messageContext
                 .getInputBundleHolder()
                 .getResource(reference.getReferenceElement())
-                .map(Observation.class::cast)
-                .get();
+                .map(Observation.class::cast);
 
-        StringBuilder stringBuilder = new StringBuilder();
+        if (observationOptional.isEmpty()) {
+            return "";
+        }
+
+        var stringBuilder = new StringBuilder();
         stringBuilder.append("{ Observation:");
 
+        var observation = observationOptional.get();
+
         if (observation.hasEffectiveDateTimeType()) {
-            stringBuilder.append(" " + formatDateUsingDayPrecision(observation.getEffectiveDateTimeType()));
+            stringBuilder
+                .append(" ")
+                .append(formatDateUsingDayPrecision(observation.getEffectiveDateTimeType()));
         } else if (observation.hasEffectivePeriod()) {
             if (observation.getEffectivePeriod().hasStart()) {
-                stringBuilder.append(" " + formatDateUsingDayPrecision(observation.getEffectivePeriod().getStartElement()));
+                stringBuilder
+                    .append(" ")
+                    .append(formatDateUsingDayPrecision(observation.getEffectivePeriod().getStartElement()));
             }
         }
 
-        CodeableConceptMappingUtils.extractTextOrCoding(observation.getCode()).ifPresent(code -> stringBuilder.append(" " + code));
+        CodeableConceptMappingUtils.extractTextOrCoding(observation.getCode()).ifPresent(
+            code -> stringBuilder
+                .append(" ")
+                .append(code)
+        );
+
         stringBuilder.append(" }");
 
         return stringBuilder.toString();
     }
 
     public static String extractReferralRequest(MessageContext messageContext, Reference reference) {
-        ReferralRequest referralRequest = messageContext
+        var referralRequestOptional = messageContext
                 .getInputBundleHolder()
                 .getResource(reference.getReferenceElement())
-                .map(ReferralRequest.class::cast)
-                .get();
+                .map(ReferralRequest.class::cast);
 
-        StringBuilder stringBuilder = new StringBuilder();
+        if (referralRequestOptional.isEmpty()) {
+            return "";
+        }
+
+        var stringBuilder = new StringBuilder();
         stringBuilder.append("{ Referral:");
 
+        var referralRequest = referralRequestOptional.get();
+
         if (referralRequest.hasAuthoredOn()) {
-            stringBuilder.append(" " + formatDateUsingDayPrecision(referralRequest.getAuthoredOnElement()));
+            stringBuilder
+                .append(" ")
+                .append(formatDateUsingDayPrecision(referralRequest.getAuthoredOnElement()));
         }
 
         if (referralRequest.hasReasonCode()) {
-            CodeableConceptMappingUtils.extractTextOrCoding(referralRequest.getReasonCode().get(0)).ifPresent(reasonCode ->
-                stringBuilder.append(" " + reasonCode)
+            CodeableConceptMappingUtils.extractTextOrCoding(referralRequest.getReasonCode().get(0)).ifPresent(
+                    reasonCode -> stringBuilder
+                        .append(" ")
+                        .append(reasonCode)
             );
         }
-
         stringBuilder.append(" }");
 
         return stringBuilder.toString();
     }
 
     public static String extractDiagnosticReport(MessageContext messageContext, Reference reference) {
-        DiagnosticReport diagnosticReport = messageContext
+        var diagnosticReportOptional = messageContext
                 .getInputBundleHolder()
                 .getResource(reference.getReferenceElement())
-                .map(DiagnosticReport.class::cast)
-                .get();
+                .map(DiagnosticReport.class::cast);
 
-        StringBuilder stringBuilder = new StringBuilder();
+        if (diagnosticReportOptional.isEmpty()) {
+            return "";
+        }
+
+        var stringBuilder = new StringBuilder();
         stringBuilder.append("{ Pathology Report:");
 
+        var diagnosticReport = diagnosticReportOptional.get();
+
         if (diagnosticReport.hasIssued()) {
-            stringBuilder.append(" " + formatDateUsingDayPrecision(diagnosticReport.getIssuedElement()));
+            stringBuilder
+                .append(" ")
+                .append(formatDateUsingDayPrecision(diagnosticReport.getIssuedElement()));
         }
 
         Optional<Identifier> identifier = diagnosticReport
-                .getIdentifier()
-                .stream()
-                .filter(d -> d.getSystem().equals(CODE_SYSTEM))
-                .findFirst();
+            .getIdentifier()
+            .stream()
+            .filter(SupportingInfoResourceExtractor::isPMIPCodeSystem)
+            .findFirst();
 
-        if (identifier.isPresent()) {
-            stringBuilder.append(" " + identifier.get().getValue());
-        }
-
+        identifier.ifPresent(
+            value -> stringBuilder
+                .append(" ")
+                .append(value.getValue())
+        );
         stringBuilder.append(" }");
 
         return stringBuilder.toString();
     }
 
     public static String extractMedicationRequest(MessageContext messageContext, Reference reference) {
-        MedicationRequest medicationRequest = messageContext
+        var medicationRequestOptional = messageContext
                 .getInputBundleHolder()
                 .getResource(reference.getReferenceElement())
-                .map(MedicationRequest.class::cast)
-                .get();
+                .map(MedicationRequest.class::cast);
 
-        StringBuilder stringBuilder = new StringBuilder();
+        if (medicationRequestOptional.isEmpty()) {
+            return "";
+        }
+
+        var stringBuilder = new StringBuilder();
         stringBuilder.append("{ Medication:");
+
+        var medicationRequest = medicationRequestOptional.get();
 
         if (medicationRequest.hasDispenseRequest()) {
             if (medicationRequest.getDispenseRequest().getValidityPeriod().hasStart()) {
-                stringBuilder.append(" "
-                        + formatDateUsingDayPrecision(
-                        medicationRequest
-                                .getDispenseRequest()
-                                .getValidityPeriod()
-                                .getStartElement()));
+                stringBuilder
+                    .append(" ")
+                    .append(
+                        formatDateUsingDayPrecision(
+                            medicationRequest.getDispenseRequest().getValidityPeriod().getStartElement()
+                        )
+                    );
             }
         }
 
         if (medicationRequest.hasMedicationReference()) {
-            Medication medication = messageContext
-                    .getInputBundleHolder()
-                    .getResource(medicationRequest.getMedicationReference().getReferenceElement())
-                    .map(Medication.class::cast)
-                    .get();
-
-            CodeableConceptMappingUtils.extractTextOrCoding(medication.getCode()).ifPresent(code ->
-                stringBuilder.append(" " + code)
-            );
+            messageContext
+                .getInputBundleHolder()
+                .getResource(medicationRequest.getMedicationReference().getReferenceElement())
+                .map(Medication.class::cast)
+                .flatMap(medication -> CodeableConceptMappingUtils.extractTextOrCoding(medication.getCode()))
+                .ifPresent(code -> stringBuilder
+                    .append(" ")
+                    .append(code)
+                );
         }
-
         stringBuilder.append(" }");
 
         return stringBuilder.toString();
@@ -173,5 +216,11 @@ public class SupportingInfoResourceExtractor {
     private static String formatDateUsingDayPrecision(BaseDateTimeType baseDateTimeType) {
         baseDateTimeType.setPrecision(TemporalPrecisionEnum.DAY);
         return DateFormatUtil.toTextFormat(baseDateTimeType);
+    }
+
+    private static boolean isPMIPCodeSystem(Identifier identifier) {
+        return StringUtils
+            .removeStart(identifier.getSystem(), URN_OID_PREFIX)
+            .equals(PMIP_CODE_SYSTEM);
     }
 }